added drop impl for texture

+++ lotus-shared/src/graphics.rs
@@ -120,15 +120,15 @@ pub mod textures {
 
     #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
     #[serde(transparent)]
-    pub struct TextureHandle(u16);
+    pub struct TextureHandle(u32);
 
     #[cfg(feature = "internals")]
     impl TextureHandle {
-        pub fn new(value: u16) -> Self {
+        pub fn new(value: u32) -> Self {
             Self(value)
         }
 
-        pub fn id(&self) -> u16 {
+        pub fn id(&self) -> u32 {
             self.0
         }
     }
